move component AvatarUser to components folder
